Setup wxWidgets sample application for windows platform

+++ include/deng/Window.h
@@ -50,14 +50,14 @@ namespace DENG {
             void glMakeCurrent();
 
             /// Toggle virtual cursor mode
-            void ToggleVCMode(TRS::Point2D<float> _origin);
+            void ToggleVCMode(TRS::Point2D<int64_t> _origin);
 
             /// Search for all required vulkan extensions
             char **FindVulkanSurfaceExtensions(uint32_t *p_ext_c) const;
             void ChangeSizeHints(neko_Hint _hints);
 
             /// Explicitly change virtual cursor mode
-            void ChangeVCMode(bool is_vcp, TRS::Point2D<float> _origin);
+            void ChangeVCMode(bool is_vcp, TRS::Point2D<int64_t> _origin);
 
             /// Change DENG cursor mode
             void ChangeCursor(neko_CursorMode cur);
